+++ CHANGELOG.md
@@ -10,6 +10,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.4.2] - 2023-10-13
+
+### Changed in 0.4.2
+
+- Changed from `int` to `int64` where required by the SenzingAPI
+- Update dependencies
+  - github.com/senzing/g2-sdk-go v0.7.3
+  - google.golang.org/grpc v1.58.3
+
+### Deleted in 0.4.2
+
+- `g2product.ValidateLicenseFile`
+- `g2product.ValidateLicenseStringBase64`
+
 ## [0.4.1] - 2023-10-03
 
 ### Changed in 0.4.1